reworking portfolio
