+++ client/components/ManagePortfolio/manage-portfolio-mobile.scss
