+++ client/router/index.js
@@ -6,9 +6,7 @@ import Profile from '@/components/Profile/Profile.vue';
 import User from '@/components/User/User.vue';
 import Signup from '@/components/Signup/Signup.vue';
 import Login from '@/components/Login/Login.vue';
-import PortfolioContainer from '@/components/Portfolio/PortfolioContainer.vue';
-import Portfolio from '@/components/Portfolio/Portfolio.vue';
-import ManagePortfolio from '@/components/Portfolio/ManagePortfolio/ManagePortfolio.vue';
+import ManagePortfolio from '@/components/ManagePortfolio/ManagePortfolio';
 import Transactions from '@/components/Transactions/Transactions';
 import Backup from '@/components/Backup/Backup';
 import BackupDone from '@/components/Backup/BackupDone';
@@ -53,20 +51,8 @@ const router = new Router({
       }
     },
     {
-      path: '/portfolio',
-      component: PortfolioContainer,
-      children: [
-        {
-          path: '/',
-          name: 'portfolio',
-          component: Portfolio
-        },
-        {
-          path: 'manage',
-          name: 'manage',
-          component: ManagePortfolio
-        }
-      ]
+      path: '/manage',
+      component: ManagePortfolio
     },
     {
       name: 'signup',
